pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opn-boss"
version = "0.1.0"
description = "OPNSense Analyzer & Recommendation Service"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    # HTTP client
    "httpx>=0.28.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    # Scheduling
    "apscheduler>=3.10.4",
    # Config & validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    # CLI
    "typer>=0.15.0",
    "rich>=13.9.0",
    # Async utilities
    "anyio>=4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "respx>=0.21.1",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
opnboss = "opn_boss.cli.commands:app"

[tool.hatch.build.targets.wheel]
packages = ["opn_boss"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
